+++ packages/pcd/pod-pcd/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./PODPCD";
 export * from "./PODPCDPackage";
-export * from "./utils";
